process_upload_function/requirements.txt
pandas==0.24.2
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-firestore==1.6.0
google-cloud-secret-manager==1.0.0
xlrd==1.0.0
gcsfs==0.6.2